QuantumStatistics.Green — Module

Provide N-body response and correlation functions

QuantumStatistics.Green.bareFermi — Method

calcualte with a given momentum vector and the chemical potential μ, rotation symmetry is assumed.

QuantumStatistics.Green.bareFermi — Method

bareFermi(β, τ, ε, [, scale])

Compute the bare fermionic Green's function. Assume $k_B=\hbar=1$

\[g(τ>0) = e^{-ετ}/(1+e^{-βε}), g(τ≤0) = -e^{-ετ}/(1+e^{βε})\]

Arguments

β: the inverse temperature

τ: the imaginary time, must be (-β, β]

ε: dispersion minus chemical potential: $E_k-μ$ it could also be the real frequency ω if the bare Green's function is used as the kernel in the Lehmann representation

QuantumStatistics.Green.bareFermiMatsubara — Method

bareFermiMatsubara(β, n, ε, [, scale])

Compute the bare Green's function for a given Matsubara frequency.

\[g(iω_n) = -1/(iω_n-ε),\]

where $ω_n=(2n+1)π/β$. The convention here is consist with the book "Quantum Many-particle Systems" by J. Negele and H. Orland, Page 95

Arguments

β: the inverse temperature

τ: the imaginary time, must be (-β, β]

ε: dispersion minus chemical potential: $E_k-μ$; it could also be the real frequency ω if the bare Green's function is used as the kernel in the Lehmann representation

QuantumStatistics.Grid.Uniform — Type

Uniform{Type,SIZE}

Create a uniform Grid with a given type and size

Member:

β: inverse temperature

halfLife: the grid is densest in the range (0, halfLife) and (β-halfLife, β)

size: the Grid size

grid: vector stores the grid

size: size of the grid vector

head: grid head

tail: grid tail

δ: distance between two grid elements

isopen: if isopen[1]==true, then grid[1] will be slightly larger than the grid head. Same for the tail.

QuantumStatistics.Grid.boseK — Function

boseK(Kf, maxK, halfLife, size::Int, kFi = floor(Int, 0.5size), twokFi = floor(Int, 2size / 3), type = Float64)

Create a logarithmic bosonic K Grid, which is densest near the momentum 0 and 2k_F

#Arguments:

Kf: Fermi momentum

maxK: the upper bound of the grid

halfLife: the grid is densest in the range (0, Kf+halfLife) and (2Kf-halfLife, 2Kf+halfLife)

size: the Grid size

kFi: index of Kf

twokFi: index of 2Kf

QuantumStatistics.Grid.fermiK — Function

fermiK(Kf, maxK, halfLife, size::Int, kFi = floor(Int, 0.5size), type = Float64)

Create a logarithmic fermionic K Grid, which is densest near the Fermi momentum $k_F$

#Arguments:

Kf: Fermi momentum

maxK: the upper bound of the grid

halfLife: the grid is densest in the range (Kf-halfLife, Kf+halfLife)

size: the Grid size

kFi: index of Kf

QuantumStatistics.Grid.tau — Function

tau(β, halfLife, size::Int, type = Float64)

Create a logarithmic Grid for the imaginary time, which is densest near the 0 and β

#Arguments:

β: inverse temperature

halfLife: the grid is densest in the range (0, halfLife) and (β-halfLife, β)

size: the Grid size

QuantumStatistics.FastMath — Module

Provide a set of fast math functions

QuantumStatistics.FastMath.invsqrt — Method

invsqrt(x)

The Legendary Fast Inverse Square Root See the following links: wikipedia and thesis

QuantumStatistics.Yeppp — Module

Forked from Yeppp.jl Please go to this link for instructions

QuantumStatistics.Yeppp.add — Method

add(x::Array, y::Array)

Perform element wise addition of the two array x and y.

QuantumStatistics.Yeppp.cos! — Method

cos(x)

Returns element wise cos of x.

QuantumStatistics.Yeppp.cos — Method

cos(x)

Returns element wise cos of x.

QuantumStatistics.Yeppp.dot — Method

dot(x::Vector{Float32}, y::Vector{Float32})

Compute the dot product of x and y.

QuantumStatistics.Yeppp.dot — Method

dot(x::Vector{Float64}, y::Vector{Float64})

Compute the dot product of x and y.

QuantumStatistics.Yeppp.dot — Method

dot(x, y)

Compute the dot product of x and y, which are two vectors with the same dimension.

QuantumStatistics.Yeppp.evalpoly — Method

evalpoly(coeff, x)

Evaluates polynomial with double precision (64-bit) floating-point coefficients coeff on an array of double precision (64-bit) floating-point elements x.

QuantumStatistics.Yeppp.exp! — Method

exp(x)

Computes the element wise exponential of x inplace.

QuantumStatistics.Yeppp.exp — Method

exp(x)

Returns the element wise exponential of x.

QuantumStatistics.Yeppp.log! — Method

log!(x)

Computes the element wise natural logarithm of x inplace.

QuantumStatistics.Yeppp.log — Method

log(x)

Returns the element wise natural logarithm of x.

QuantumStatistics.Yeppp.max — Method

max(x, y)

Compares the vectors x and y and return the element wise maxima.

QuantumStatistics.Yeppp.min — Method

max(x, y)

Compares the vectors x and y and return the element wise minima.

QuantumStatistics.Yeppp.multiply — Method

multiply(x, y)

Perform element wise multiplication of the two array x and y.

QuantumStatistics.Yeppp.sin! — Method

sin(x)

Computes element wise sin of x inplace.

QuantumStatistics.Yeppp.sin — Method

sin(x)

Returns element wise sin of x.

QuantumStatistics.Yeppp.subtract — Method

subtract(x::Array, y::Array)

Perform element wise subtraction of the two array x and y.

QuantumStatistics.Yeppp.tan! — Method

tan(x)

Returns element wise tan of x.

QuantumStatistics.Yeppp.tan — Method

tan(x)

Returns element wise tan of x.