Monte Carlo

QuantumStatistics.MonteCarlo — Module

utility for Monte Carlo

QuantumStatistics.MonteCarlo.createFermiK! — Function

createFermiK!(newK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.createIdx! — Function

createIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.createT! — Method

createT!(newT, β=1.0, rng=GLOBAL_RNG)

Propose to generate new tau (uniformly) randomly in [0, β)

Arguments

newT: index ∈ [0, β)

β=1.0: inverse temperature

QuantumStatistics.MonteCarlo.removeFermiK — Function

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.removeIdx — Function

removeIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.removeT — Method

removeT(oldT, β=1.0, rng=GLOBAL_RNG)

Propose to remove old tau in [0, β)

QuantumStatistics.MonteCarlo.shiftIdx! — Function

shiftIdx!(oldIdx::Int, newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.shiftK! — Function

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector

QuantumStatistics.MonteCarlo.shiftK_flip! — Method

shiftK_flip!(oldK, newK)

Propose to flip oldK to newK. Work for generic momentum vector

QuantumStatistics.MonteCarlo.shiftK_radial! — Function

shiftK_radial!(oldK, newK, λ=1.5, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector

Arguments

newK: randomly proposed in [oldK/λ, oldK*λ)

QuantumStatistics.MonteCarlo.shiftT! — Method

shiftT!(oldT, newT, β=1.0, rng=GLOBAL_RNG)

Propose to shift the old tau to new tau, both in [0, β)

Arguments

newT: will be modified!

QuantumStatistics.MonteCarlo.shiftT_flip! — Method

shiftT_flip!(oldT, newT, β=1.0, rng=GLOBAL_RNG)

Propose to flip the old tau to a new tau, both in [0, β)

Arguments

newT: will be modified!