Monte Carlo

QuantumStatistics.MonteCarlo — Module

Monte Carlo Calculator for Diagrams

QuantumStatistics.MonteCarlo.Configuration — Type

mutable struct Configuration

Struct that contains everything needed for MC.

There are three different componenets:

Members

Static parameters

seed: seed to initialize random numebr generator, also serves as the unique pid of the configuration

rng: a MersenneTwister random number generator, seeded by seed

para: user-defined parameter, could be set to nothing if not needed

totalStep: the total number of updates for this configuration

var: TUPLE of variables, each variable should be derived from the abstract type Variable, see variable.jl for details). Use a tuple rather than a vector improves the performance.

integrand properties

neighbor::Vector{Vector{Int}} : vectors that indicates the neighbors of each integrand. e.g., ([2, ], [1, ]) means the neighbor of the first integrand is the second one, while the neighbor of the second integrand is the first. There is a MC update proposes to jump from one integrand to another. If these two integrands' degrees of freedom are very different, then the update is unlikely to be accepted. To avoid this problem, one can specify neighbor to guide the update.

By default, we assume the N integrands are in the increase order, meaning the neighbor will be set to ([N+1, 2], [1, 3], [2, 4], ..., [N-1,], [1, ]), where the first N entries are for diagram 1, 2, ..., N and the last entry is for the normalization diagram. Only the first diagram is connected to the normalization diagram.

dof::Vector{Vector{Int}}: degrees of freedom of each integrand, e.g., [[0, 1], [2, 3]] means the first integrand has zero var#1 and one var#2; while the second integrand has two var#1 and 3 var#2.

observable: observables that is required to calculate the integrands, will be used in the measure function call. It is either an array of any type with the common operations like +-*/^ defined.

reweight: reweight factors for each integrands.

visited: how many times this integrand is visited by the Markov chain.

current MC state

step: the number of MC updates performed up to now

curr: the current integrand, initialize with 1

norm: the index of the normalization diagram. norm is larger than the index of any user-defined integrands

absWeight: the abolute weight of the current integrand. User is responsible to initialize it after the contructor is called.

normalization: the accumulated normalization factor. Physical observable = Configuration.observable/Configuration.normalization.

propose/accept: array to store the proposed and accepted updates for each integrands and variables. Their shapes are (number of updates X integrand number X max(integrand number, variable number). The last index will waste some memory, but the dimension is small anyway.

QuantumStatistics.MonteCarlo.create! — Function

createIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.create! — Function

create!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to generate a new pair of tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: TauPair variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Function

create!(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to generate new angle (uniformly) randomly in [0, 2π), return proposal probability

Arguments

theta: angle variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Function

create!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to generate new tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Method

create!(K::FermiK{D}, idx::Int, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.remove — Function

removeIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.remove — Function

remove(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to remove an existing pair of tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Function

remove(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to remove old theta in [0, 2π), return proposal probability

Arguments

- `theta`:  Tau variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Function

remove(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to remove old tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Method

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.sample — Method

sample(config::Configuration, integrand::Function, measure::Function; Nblock=16, print=0, printio=stdout, save=0, saveio=nothing, timer=[])

sample the integrands, collect statistics, and return the expected values and errors.

Remarks

User may run the MC in parallel using MPI. Simply run mpiexec -n N julia userscript.jl where N is the number of workers. In this mode, only the root process returns meaningful results. All other workers return nothing, nothing. User is responsible to handle the returning results properly.

In the MC, a normalization diagram is introduced to normalize the MC estimates of the integrands. More information can be found in the link: https://kunyuan.github.io/QuantumStatistics.jl/dev/man/important_sampling/#Important-Sampling. User don't need to explicitly specify this normalization diagram.Internally, normalization diagram will be added to each table that is related to the integrands.

Arguments

config: Configuration struct

integrand: function call to evaluate the integrand. It should accept an argument of the type Configuration, and return a weight. Internally, MC only samples the absolute value of the weight. Therefore, it is also important to define Main.abs for the weight if its type is user-defined.

measure: function call to measure. It should accept an argument of the type Configuration, then manipulate observables obs.

Nblock: repeat times. The tasks will automatically distributed to multi-process in MPI mode.

print: -1 to not print anything, 0 to print minimal information, >0 to print summary for every print seconds

printio: io to print the information

save: -1 to not save anything, 0 to save observables obs in the end of sampling, >0 to save observables obs for every save seconds

saveio: io to save

QuantumStatistics.MonteCarlo.shift! — Function

shiftIdx!(oldIdx::Int, newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.shift! — Function

shift!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau pair to a new tau pair, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Function

shift!(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to shift the old theta to new theta, both in [0, 2π), return proposal probability

Arguments

theta: angle variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Function

shift!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau to a new tau, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Method

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector