Monte Carlo

QuantumStatistics.MonteCarlo — Module

Monte Carlo Calculator for Diagrams

QuantumStatistics.MonteCarlo.Diagram — Type

Group{A}(type::Int, internal::Tuple{Vararg{Int}}, external::Tuple{Vararg{Int}}, eval, obstype=Float64)

create a group of diagrams

#Arguments:

type: integer identifier of the group

internal: internal variable numbers, e.g. [number of internal momentum, number of internal tau]

external: array of size of external index, e.g. [size of external momentum index, size of external tau]

eval: function to evaluate the group

obstype: type of the diagram weight, e.g. Float64

QuantumStatistics.MonteCarlo.create! — Function

createIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.create! — Function

create!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to generate new tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Method

create!(K::FermiK{D}, idx::Int, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.progressBar — Method

progressBar(step, total)

Return string of progressBar (step/total*100%)

QuantumStatistics.MonteCarlo.remove — Function

removeIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.remove — Function

remove(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to remove old tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Method

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.shift! — Function

shiftIdx!(oldIdx::Int, newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.shift! — Function

shift!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to shift the old tau to new tau, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Method

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector