Monte Carlo

QuantumStatistics.MonteCarlo — Module

Monte Carlo Calculator for Diagrams

QuantumStatistics.MonteCarlo.Configuration — Type

mutable struct Configuration

Struct that saves everything needed by MC.

There are three different pieces of information:

Members

Static parameters

seed: seed to initialize random numebr generator, also serves as the unique pid of the configuration

rng: a MersenneTwister random number generator, seeded by seed

para: user-defined parameter, could be nothing if not needed

totalStep: the total number of updates for this configuration

var: TUPLE of variables, each variable should be derived from the abstract type Variable, see variable.jl for details). Use a tuple rather than a vector improves the performance.

integrand properties

neighbor: vectors that indicates the neighbors of each integrand. e.g., ([2, ], [1, ]) means the neighbor of the first integrand is the second one, while the neighbor of the second integrand is the first. There is a MC update proposes to jump from one integrand to another. If these two integrands' degrees of freedom are very different, then the update is unlikely to be accepted. To avoid this problem, one can specify neighbor to guide the update. By default, we assume the N integrands are in the increase order, meaning the neighbor will be set to ([2, ], [1, 3], [2, 4], ..., [N-1,])

dof: degrees of freedom of each integrand, e.g., ([0, 1], [2, 3]) means the first integrand has zero var#1 and one var#2; while the second integrand has two var#1 and 3 var#2.

observable: observables that is required to calculate the integrands, will be used in the measure function call

reweight: reweight factors for each integrands. If not set, then all factors will be initialized as one.

visited: how many times this integrand is visited by the Markov chain.

current MC state

step: the number of MC updates performed up to now

curr: the current integrand

absWeight: the abolute weight of the current integrand

propose/accept: array to store the proposed and accepted updates for each integrands and variables. Their shapes are (number of updates X integrand number X max(integrand number, variable number). The last index will waste some memory, but the dimension is small anyway.

QuantumStatistics.MonteCarlo.create! — Function

create!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to generate new tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Function

createIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.create! — Function

create!(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to generate new angle (uniformly) randomly in [0, 2π), return proposal probability

Arguments

theta: angle variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Function

create!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to generate a new pair of tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: TauPair variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.create! — Method

create!(K::FermiK{D}, idx::Int, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.progressBar — Method

progressBar(step, total)

Return string of progressBar (step/total*100%)

QuantumStatistics.MonteCarlo.remove — Function

remove(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to remove old tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Function

removeIdx!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.remove — Function

remove(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to remove old theta in [0, 2π), return proposal probability

Arguments

- `theta`:  Tau variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Function

remove(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to remove an existing pair of tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.remove — Method

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

QuantumStatistics.MonteCarlo.sample — Function

sample(totalStep, var, dof, obs, integrand::Function, measure::Function, normalize::Function=nothing; Nblock=16, para=nothing, neighbor=nothing, seed=nothing, reweight=nothing, print=0, printio=stdout, save=0, saveio=nothing, timer=[])

sample the integrands, collect statistics, and return the expected values and errors

Arguments

totalStep: total number of updates

var: TUPLE of variables, each variable should be derived from the abstract type Variable, see variable.jl for details). Use a tuple rather than a vector improves the performance.

dof: degrees of freedom of each integrand, e.g., ([0, 1], [2, 3]) means the first integrand has zero var#1 and one var#2; while the second integrand has two var#1 and 3 var#2.

obs: observables that is required to calculate the integrands, will be used in the measure function call

integrand: function call to evaluate the integrand. It should accept an argument of the type Configuration, and return a weight. Internally, MC only samples the absolute value of the weight. Therefore, it is also important to define Main.abs for the weight if its type is user-defined.

measure: function call to measure. It should accept an argument of the type Configuration, then manipulate observables obs.

normalize: function call to derive averages and errors from observables obs. It should accept an argument of the type Configuration.

Nblock: repeat times. The tasks will automatically distributed to multi-process if Distributed package is imported globally.

para: user-defined parameter that is useful in the functions integrand, measure, normalize.

neighbor: vectors that indicates the neighbors of each integrand. e.g., ([2, ], [1, ]) means the neighbor of the first integrand is the second one, while the neighbor of the second integrand is the first. There is a MC update proposes to jump from one integrand to another. If these two integrands' degrees of freedom are very different, then the update is unlikely to be accepted. To avoid this problem, one can specify neighbor to guide the update. By default, we assume the N integrands are in the increase order, meaning the neighbor will be set to ([2, ], [1, 3], [2, 4], ..., [N-1,])

seed: the seed for random number generator. If seed is set, then the random number generator of each blocks will be initialized with seed+1, seed+2, .... On the other hand, if seed is nothing, then MC will call RandomDevice() to get a system-generated seed for each block. Since seed is different for each block, it could also be used as the id of each block.

reweight: reweight factors for each integrands. If not set, then all factors will be initialized as one.

print: -1 to not print anything, 0 to print minimal information, >0 to print summary for every print seconds

printio: io to print the information

save: -1 to not save anything, 0 to save observables obs in the end of sampling, >0 to save observables obs for every save seconds

saveio: io to save

QuantumStatistics.MonteCarlo.shift! — Function

shift!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau to a new tau, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Function

shiftIdx!(oldIdx::Int, newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

QuantumStatistics.MonteCarlo.shift! — Function

shift!(theta::Angle, idx::Int, rng=GLOBAL_RNG)

Propose to shift the old theta to new theta, both in [0, 2π), return proposal probability

Arguments

theta: angle variable

idx: theta.t[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Function

shift!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau pair to a new tau pair, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

QuantumStatistics.MonteCarlo.shift! — Method

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector